+++ src/fea/fem/tests/test_hex8.py
@@ -11,7 +11,7 @@ import taichi as ti
 @pytest.fixture(scope="module", autouse=True)
 def init_taichi():
     """모듈당 한 번 Taichi 초기화."""
-    ti.init(arch=ti.cpu, default_fp=ti.f32)
+    ti.init(arch=ti.cpu, default_fp=ti.f64)
 
 
 class TestHEX8ShapeFunctions:
@@ -107,7 +107,7 @@ class TestHEX8Mesh:
             [1.0, 0.0, 1.0],  # 5
             [1.0, 1.0, 1.0],  # 6
             [0.0, 1.0, 1.0],  # 7
-        ], dtype=np.float32)
+        ], dtype=np.float64)
 
         elements = np.array([[0, 1, 2, 3, 4, 5, 6, 7]], dtype=np.int32)
 
@@ -139,7 +139,7 @@ class TestHEX8Mesh:
             [L, 0.0, H],
             [L, W, H],
             [0.0, W, H],
-        ], dtype=np.float32)
+        ], dtype=np.float64)
 
         elements = np.array([[0, 1, 2, 3, 4, 5, 6, 7]], dtype=np.int32)
 
@@ -170,7 +170,7 @@ class TestHEX8Mesh:
             [1.0, 0.0, 1.0],
             [1.0, 1.0, 1.0],
             [0.0, 1.0, 1.0],
-        ], dtype=np.float32)
+        ], dtype=np.float64)
 
         elements = np.array([[0, 1, 2, 3, 4, 5, 6, 7]], dtype=np.int32)
 
@@ -210,7 +210,7 @@ class TestHEX8Solver:
             for j in range(ny+1):
                 for i in range(nx+1):
                     nodes.append([i * spacing, j * spacing, k * spacing])
-        nodes = np.array(nodes, dtype=np.float32)
+        nodes = np.array(nodes, dtype=np.float64)
 
         # 요소 연결성
         elements = []
@@ -238,7 +238,7 @@ class TestHEX8Solver:
 
         # 상단면에 압축력
         top_nodes = np.where(nodes[:, 2] > nz*spacing - 0.1)[0]
-        forces = np.zeros((len(top_nodes), 3), dtype=np.float32)
+        forces = np.zeros((len(top_nodes), 3), dtype=np.float64)
         forces[:, 2] = -100.0  # -z 방향
         mesh.set_nodal_forces(top_nodes, forces)
